- Removed all compiler warnings in the code

+++ RPiAPMCopter_upstream/BattMonitor.cpp
@@ -20,7 +20,7 @@ void BattMonitor::setup_source(int_fast8_t volt_pin,
   init();
 }
 
-void BattMonitor::setup_source(const uint_fast8_t &t) {
+void BattMonitor::setup_source(const int &t) {
   switch (t) {
     case ATTO45:
       setup_source(AP_ATTO_VOLT_PIN, AP_ATTO_CURR_PIN, AP_BATT_VOLTDIVIDER_ATTO45, AP_BATT_CURR_AMP_PERVOLT_ATTO45, AP_BATT_CAPACITY_DEFAULT, 0, AP_BATT_MONITOR_VOLTAGE_AND_CURRENT);
